// src/app/models/sales/ILogSearchParameters.ts
export interface ILogSearchParameters {
    first: number;
    PageNo: number;
    pageCount: number;
    rows: number;
    limit: number;
    sortOrder: number;
    sortField: number;
    columnFilters: {};
    filters: any;
    globalFilter: string;
    exception: string;
    message: string;
    messageTemplate: string;
    properties: string;
    level: string;
    timestamp: string;
}